TODO process monitor game

+++ src/Aueio/ClubBundle/Controller/PlayerController.php
@@ -25,14 +25,15 @@ class PlayerController extends Controller
 	*/
 	public function viewAction($id)
     {
-    	$player = $this->getDoctrine()
-				    	->getRepository('AueioClubBundle:Player')
-				    	->find($id);
+    	$em = $this->getDoctrine()->getEntityManager();
+    	$player = $em->getRepository('AueioClubBundle:Player')->find($id);
     	if (!$player) {
     		throw $this->createNotFoundException('No player found for id '.$id);
     	}
-    	
-    	return $this->render('AueioClubBundle:Player:view.html.twig', array('player' => $player, 'me' => false));
+		$stats = $em->getRepository('AueioClubBundle:Action')->getStats($id);
+		$stats['total'] = count($player->getTeam()->getRoles());
+		
+    	return $this->render('AueioClubBundle:Player:view.html.twig', array('player' => $player, 'stats' =>$stats));
     }
     /**
      * @Route("/list")
@@ -52,10 +53,14 @@ class PlayerController extends Controller
     	if (!$player) {
     		throw $this->createNotFoundException('No player found for id '.$id);
     	}
-    	 
-    	$em->remove($player);
-    	$em->flush();
-    	 
-    	return $this->redirect($this->generateUrl('aueio_club_player_list'));
+    	if ( ($player->getId() == $this->get('security.context')->getToken()->getUser()->getId())
+    			|| $this->get('security.context')->isGranted('ROLE_ADMIN')) {
+	    	$em->remove($player);
+	    	$em->flush();
+	    	 
+	    	return $this->redirect($this->generateUrl('fos_user_security_logout'));
+    	}else{
+    		return $this->redirect($this->get('request')->headers->get('referer'));
+    	}
     }
 }
